sample/Tests/requirements-mobile.txt
AltTester_Driver==2.1.1
google_api_python_client==2.136.0
google_auth_oauthlib==1.2.0
protobuf==5.27.2
selenium==4.22.0
pytest==8.2.2
requests==2.32.3
mailslurp-client==15.19.22
Appium-Python-Client
browserstack-sdk
browserstack-local
pytest-variables
pytest-selenium